Create Texture Collector v4.5.lua
This version doesn't rely on the png library. It instead makes a compressed image file that only takes 3 bytes per rgb color. It also uses fastlz to compress the result which helps out a ton in the case of maps like bigcity that have very predictable textures with bland windows and stuff.